+++ ROUTES/tableRoute.js
@@ -3,18 +3,24 @@ const router = express.Router();
 const {Table} = require ('../MODELS/tables');
 
 router.post(`/`,  async (req, res) => {
-    let table = new Table({
-        name: req.body.name,
-        guests: req.body.guests
-    })
-    table = await table.save();
-    if(!table){
-        return res.status(400).send('order cant be created. ');
-    }
-    res.send(table);
+        let table = new Table({
+            name: req.body.name,
+            guests: req.body.guests
+        })
+        if(req.body.name == null || req.body.guests == null){
+            return res.status(400).send('client error, cannot be created')
+        }
+        table = await table.save();
+        if(!table){
+            return res.status(400).send('order cant be created. ');
+        }
+        res.status(200).send(table);
     
 });
 router.get(`/:id`, async(req, res) => {
+    if(! req.params.id.match(/^[0-9a-fA-F]{24}$/)){
+        return res.status(400).send('BAD CLIENT REQUEST, invalid object Id');
+    }
     const table = await Table.findById(req.params.id);
     if(table){
         return res.status(200).send(table)
@@ -24,6 +30,9 @@ router.get(`/:id`, async(req, res) => {
 });
 
 router.put(`/:id`, async (req, res) => {
+    if(! req.params.id.match(/^[0-9a-fA-F]{24}$/)){
+        return res.status(400).send('BAD CLIENT REQUEST, invalid object Id');
+    }
     const table = await Table.findByIdAndUpdate(
         req.params.id,
         {
@@ -33,11 +42,14 @@ router.put(`/:id`, async (req, res) => {
         {new: true}
     )
     if(!table){
-        res.status(500).json({message: 'Table id not recognized. Table connot be updated '})
+        res.status(400).json({message: 'Table id not recognized. Table connot be updated '})
     }
     res.status(200).send(table);
 })
 router.delete(`/:id`, (req, res)=> {
+    if(! req.params.id.match(/^[0-9a-fA-F]{24}$/)){
+        return res.status(400).send('BAD CLIENT REQUEST, invalid object Id');
+    }
     Table.findByIdAndRemove(req.params.id).then(table => {
         if(table){
             return res.status(200).json({success: true, message: 'table successfully deleted!'})
